face_api/requirements.txt
face_recognition
flask
flask-cors
opencv-python
numpy
pandas
openpyxl
Pillow